Refactor of real, rectangular and polar interval classes
The properties are now stored as a vector of the given class. Some of the operations have been updated to be more efficient using vector operations. The parenthesis operators have been overloaded so that the syntax stays the same for retrocompatibility.

Be aware of the issue : the previous syntax [obj.Infimum] can now return a transposed version of what it would have returned.