Add files via upload
Bugfix: "I found the bug with the backtracking; if vertices in the polygon are duplicated, the outer angle is not well defined, and it fooled some of the conditions. I added a "v = unique(v, 'stable')" to the backtracking, which fixes the issue. I also made the code a bit more robust by considering the edge cases more carefully, so I also removed all eps10 terms since they should not be needed."